Conflicts:
	app/app.iml
	app/build.gradle
	app/src/main/AndroidManifest.xml
	app/src/main/java/com/mamba/grapple/Chat.java
	app/src/main/java/com/mamba/grapple/DBService.java
	app/src/main/java/com/mamba/grapple/LocationObject.java
	app/src/main/java/com/mamba/grapple/Results.java
	app/src/main/java/com/mamba/grapple/Search.java
	app/src/main/java/com/mamba/grapple/Splash.java
	app/src/main/java/com/mamba/grapple/Tutor.java
	app/src/main/java/com/mamba/grapple/TutorObject.java
	app/src/main/java/com/mamba/grapple/TutorsAdapter.java
	app/src/main/res/layout/activity_addresslist.xml
	app/src/main/res/layout/activity_register.xml
	app/src/main/res/layout/activity_search.xml
	app/src/main/res/layout/activity_signin.xml
	app/src/main/res/layout/activity_splash.xml
	app/src/main/res/layout/activity_tutor.xml
	app/src/main/res/layout/activity_tutorchat.xml
	app/src/main/res/layout/activity_tutorselect.xml
	app/src/main/res/layout/chat_message_left.xml
	app/src/main/res/layout/tutor_row.xml

+++ app/src/main/java/com/mamba/grapple/TutorLocation.java
@@ -9,12 +9,12 @@ import android.os.Parcelable;
 
 
 public class TutorLocation implements Parcelable {
-    public float xPos;
-    public float yPos;
+    public double xPos;
+    public double yPos;
 
     protected TutorLocation(Parcel in) {
-        xPos = in.readFloat();
-        yPos = in.readFloat();
+        xPos = in.readDouble();
+        yPos = in.readDouble();
     }
 
     @Override
@@ -24,8 +24,8 @@ public class TutorLocation implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeFloat(xPos);
-        dest.writeFloat(yPos);
+        dest.writeDouble(xPos);
+        dest.writeDouble(yPos);
     }
